remove the dreaming feature
Remove the notebook dreaming/harvest feature end-to-end. The nightly
harvest_dream task accumulated a deduplicated candidate set under
.attn/dreams/ but had no downstream consumer: the promote pass that would
turn candidates into durable memory was never built, and memory notes are
authored by hand. The feature was gated off by default and never shipped to
users, so this removes it rather than carrying dead surface area.

Removed:
- harvest_dream task kind + executor and its durable-runner registration
- the dreaming candidate store and run-state (internal/notebook/dreams*.go)
- notebook_dream_status / notebook_dream_run protocol messages + results
- the `attn notebook dream` CLI and client methods
- the notebook.dreaming.enabled setting

Kept (surgically): the per-minute notebook cron now drives only the daily
per-workspace narrate backstop. Its schedule/timezone still live under the
notebook.dreaming.* setting keys and the dreaming*-named helpers in
notebook_dreaming_scheduler.go for now; a follow-up renames that cron infra
to notebook.cron.* so no "dreaming" identifier survives.

Protocol bumped to 111 (removed commands/messages force a reconnect).

Verified: go build/vet, go test ./internal/... ./cmd/... (daemon scoped to
dodge the gitstatus race), frontend tsc + vitest (949).
