+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Format: `[YYYY-MM-DD]` entries with categories: Added, Changed, Fixed, Removed.
 ## [2026-06-15]
 
 ### Added
-- **See what attn's background work is doing, from the app or `attn notebook tasks`.** A new Tasks section in the Notebook browser — and a matching command — lists attn's durable background tasks — the context-compaction, session-digest, journal-narration, and nightly memory-harvest jobs — showing each one's state, what it's working on, how many attempts it has taken, when it next runs, and the last error if it failed. The panel refreshes live as tasks change, and a stuck (failed or dead) task gets a Retry button to run it again now. It's a window into work that used to be invisible, so you can tell at a glance whether something is queued, running, retrying, or stuck.
+- **See what attn's background work is doing, from the app or `attn notebook tasks`.** A new Tasks section in the Notebook browser — and a matching command — lists attn's durable background tasks — the context-compaction, session-digest, and journal-narration jobs — showing each one's state, what it's working on, how many attempts it has taken, when it next runs, and the last error if it failed. The panel refreshes live as tasks change, and a stuck (failed or dead) task gets a Retry button to run it again now. It's a window into work that used to be invisible, so you can tell at a glance whether something is queued, running, retrying, or stuck.
 - **Your work-journal now writes itself from the work your agents actually do.** As sessions finish, attn quietly digests each one, and when a workspace is active — or the moment it's removed — it composes a curated, dated journal entry for that workspace: the real decisions and why, the fights and how they resolved, what shipped, the dead-ends, and what's still open. It grounds the story in what actually happened (passing tests, real commits) rather than what an agent claimed, surfaces silent wins and abandoned plans, and continues the narrative across days instead of repeating itself. A removed workspace gets a final retrospective so the whole effort is captured before it's gone. Entries are written safely, one per workspace per day, and never include secrets.
 - **Active long-lived workspaces now get a daily journal entry even when no session ends that day.** A nightly pass writes a fresh entry for any workspace you actually worked in that day — a session finished or its shared context changed — so a workspace that runs for weeks stays journaled day to day instead of only being written up when it's removed. Workspaces you didn't touch are skipped, so the pass never spends effort on idle ones.
 
@@ -27,7 +27,6 @@ Format: `[YYYY-MM-DD]` entries with categories: Added, Changed, Fixed, Removed.
 ### Added
 - **Every agent now helps keep your work-journal, not just the chief of staff.** Whenever attn launches an agent in a workspace, it now teaches it to jot a short journal entry when something is worth remembering — a decision and why, a hard-won fix, something meaningful it finished, or a durable gotcha — so your journal reflects work across the whole workspace, not only delegated chief-of-staff dispatches. Entries stay short and factual, are written through attn (so they're safe and observable), and never include secrets.
 - **Your Notebook journal now writes itself when delegated work finishes.** When a chief-of-staff dispatch completes or fails — or its session simply ends — attn automatically records it in that day's journal: a dated, human-readable entry with what was done, any decision that was reached, and how it was verified. You get a durable history of what was built and decided without anyone having to remember to write it down, and each entry links back to the dispatch it came from. Entries are recorded once, even if attn restarts.
-- **Preview what the Notebook's nightly memory pass would keep.** A new `attn notebook dream` command shows the recurring, durable facts attn could distill into lasting memory — gathered from your journals, your workspaces' decisions and constraints, and finished chief-of-staff dispatches, with the ones that recur across several places ranked first. `attn notebook dream status` is a quick summary; `attn notebook dream --dry-run` lists the candidates. It's read-only — nothing is written yet — so you can see the raw material before the automatic consolidation pass that acts on it arrives in an upcoming release.
 
 ### Changed
 - **Automatic workspace-context compaction is now durable across restarts.** When a workspace's shared context grows large enough to compact, the pending compaction is now queued on attn's durable task engine instead of an in-memory timer, so it survives a daemon restart and is no longer lost if attn closes during the wait. The compaction agent now works with ordinary file tools instead of a constrained tool pair; what you see is unchanged — oversized context still gets compacted, validated, and applied safely, and is cancelled cleanly when its workspace goes away.
