+++ cmd/attn/main_test.go
@@ -521,59 +521,6 @@ func TestParseNotebookJournalArgs(t *testing.T) {
 	}
 }
 
-func TestPrintDreamStatus(t *testing.T) {
-	var buf bytes.Buffer
-	printDreamStatus(&buf, &protocol.NotebookDreamStatusResult{
-		Enabled:           false,
-		CandidateCount:    3,
-		MultiContextCount: 1,
-		PersistedCount:    2,
-		Schedule:          protocol.Ptr("0 3 * * *"),
-		Timezone:          protocol.Ptr("America/New_York"),
-		NextRunAt:         protocol.Ptr("2026-06-15T07:00:00Z"),
-		LastRunAt:         protocol.Ptr("2026-06-14T07:00:00Z"),
-		SourceCounts: []protocol.NotebookDreamSourceCount{
-			{Source: "context", Count: 1},
-			{Source: "journal", Count: 2},
-		},
-		Top: []protocol.NotebookDreamCandidate{
-			{Source: "context", Title: protocol.Ptr("ws-a"), Snippet: "Daemon owns every notebook write.", Occurrences: 2, Contexts: []string{"workspace:ws-a", "workspace:ws-b"}},
-		},
-	})
-	out := buf.String()
-	for _, want := range []string{
-		"dreaming: disabled",
-		"schedule: 0 3 * * * (America/New_York), next run ",
-		"last run: ",
-		"candidates: 3 (1 across multiple contexts), 2 persisted",
-		"context:", "journal:", "[2× ·2 ctx]", "Daemon owns every notebook write.",
-	} {
-		if !strings.Contains(out, want) {
-			t.Fatalf("status output missing %q:\n%s", want, out)
-		}
-	}
-}
-
-func TestPrintDreamRunPreviewOnly(t *testing.T) {
-	var buf bytes.Buffer
-	printDreamRun(&buf, &protocol.NotebookDreamRunResult{
-		Applied:           false,
-		CandidateCount:    1,
-		MultiContextCount: 0,
-		SourceCounts:      []protocol.NotebookDreamSourceCount{{Source: "journal", Count: 1}},
-		Candidates: []protocol.NotebookDreamCandidate{
-			{Source: "journal", Snippet: "A durable fact.", Occurrences: 1, Contexts: []string{"journal:2026-06-13"}},
-		},
-	})
-	out := buf.String()
-	if !strings.Contains(out, "preview only, nothing written") {
-		t.Fatalf("run output should state it is preview-only:\n%s", out)
-	}
-	if !strings.Contains(out, "A durable fact.") {
-		t.Fatalf("run output missing the candidate snippet:\n%s", out)
-	}
-}
-
 func TestParseNotebookMemoryArgs(t *testing.T) {
 	t.Run("create (no base-hash, stdin)", func(t *testing.T) {
 		got, err := parseNotebookMemoryArgs([]string{"write", "--path", "/memory/decisions/x.md"})
