+++ internal/protocol/constants.go
@@ -10,7 +10,7 @@ import (
 // ProtocolVersion is the version of the daemon-client protocol.
 // Increment this when making breaking changes to the protocol.
 // Client and daemon must have matching versions.
-const ProtocolVersion = "110"
+const ProtocolVersion = "111"
 
 // CapabilityWorkspaceSessions is required for websocket clients that use the
 // interactive daemon API. Clients without it are not workspace-first clients.
@@ -66,8 +66,6 @@ const (
 	CmdNotebookGuide                      = "notebook_guide"
 	CmdNotebookBacklinks                  = "notebook_backlinks"
 	CmdNotebookSendToChief                = "notebook_send_to_chief"
-	CmdNotebookDreamStatus                = "notebook_dream_status"
-	CmdNotebookDreamRun                   = "notebook_dream_run"
 	CmdNotebookTaskList                   = "notebook_task_list"
 	CmdNotebookTaskRetry                  = "notebook_task_retry"
 	CmdUnregister                         = "unregister"
@@ -506,20 +504,6 @@ func ParseMessage(data []byte) (string, interface{}, error) {
 
 	case CmdNotebookSendToChief:
 		var msg NotebookSendToChiefMessage
-		if err := json.Unmarshal(data, &msg); err != nil {
-			return "", nil, err
-		}
-		return peek.Cmd, &msg, nil
-
-	case CmdNotebookDreamStatus:
-		var msg NotebookDreamStatusMessage
-		if err := json.Unmarshal(data, &msg); err != nil {
-			return "", nil, err
-		}
-		return peek.Cmd, &msg, nil
-
-	case CmdNotebookDreamRun:
-		var msg NotebookDreamRunMessage
 		if err := json.Unmarshal(data, &msg); err != nil {
 			return "", nil, err
 		}
