+++ internal/protocol/generated.go
@@ -1758,105 +1758,6 @@ type NotebookChangedMessage struct {
 	Paths []string `json:"paths"`
 }
 
-type NotebookDreamCandidate struct {
-	// Contexts corresponds to the JSON schema field "contexts".
-	Contexts []string `json:"contexts"`
-
-	// FirstSeen corresponds to the JSON schema field "first_seen".
-	FirstSeen *string `json:"first_seen,omitempty,omitzero"`
-
-	// LastSeen corresponds to the JSON schema field "last_seen".
-	LastSeen *string `json:"last_seen,omitempty,omitzero"`
-
-	// Occurrences corresponds to the JSON schema field "occurrences".
-	Occurrences int `json:"occurrences"`
-
-	// SignalKey corresponds to the JSON schema field "signal_key".
-	SignalKey string `json:"signal_key"`
-
-	// Snippet corresponds to the JSON schema field "snippet".
-	Snippet string `json:"snippet"`
-
-	// Source corresponds to the JSON schema field "source".
-	Source string `json:"source"`
-
-	// Sources corresponds to the JSON schema field "sources".
-	Sources []string `json:"sources"`
-
-	// Title corresponds to the JSON schema field "title".
-	Title *string `json:"title,omitempty,omitzero"`
-}
-
-type NotebookDreamRunMessage struct {
-	// Apply corresponds to the JSON schema field "apply".
-	Apply *bool `json:"apply,omitempty,omitzero"`
-
-	// Cmd corresponds to the JSON schema field "cmd".
-	Cmd string `json:"cmd"`
-}
-
-type NotebookDreamRunResult struct {
-	// Applied corresponds to the JSON schema field "applied".
-	Applied bool `json:"applied"`
-
-	// CandidateCount corresponds to the JSON schema field "candidate_count".
-	CandidateCount int `json:"candidate_count"`
-
-	// Candidates corresponds to the JSON schema field "candidates".
-	Candidates []NotebookDreamCandidate `json:"candidates"`
-
-	// MultiContextCount corresponds to the JSON schema field "multi_context_count".
-	MultiContextCount int `json:"multi_context_count"`
-
-	// SourceCounts corresponds to the JSON schema field "source_counts".
-	SourceCounts []NotebookDreamSourceCount `json:"source_counts"`
-}
-
-type NotebookDreamSourceCount struct {
-	// Count corresponds to the JSON schema field "count".
-	Count int `json:"count"`
-
-	// Source corresponds to the JSON schema field "source".
-	Source string `json:"source"`
-}
-
-type NotebookDreamStatusMessage struct {
-	// Cmd corresponds to the JSON schema field "cmd".
-	Cmd string `json:"cmd"`
-}
-
-type NotebookDreamStatusResult struct {
-	// CandidateCount corresponds to the JSON schema field "candidate_count".
-	CandidateCount int `json:"candidate_count"`
-
-	// Enabled corresponds to the JSON schema field "enabled".
-	Enabled bool `json:"enabled"`
-
-	// LastRunAt corresponds to the JSON schema field "last_run_at".
-	LastRunAt *string `json:"last_run_at,omitempty,omitzero"`
-
-	// MultiContextCount corresponds to the JSON schema field "multi_context_count".
-	MultiContextCount int `json:"multi_context_count"`
-
-	// NextRunAt corresponds to the JSON schema field "next_run_at".
-	NextRunAt *string `json:"next_run_at,omitempty,omitzero"`
-
-	// PersistedCount corresponds to the JSON schema field "persisted_count".
-	PersistedCount int `json:"persisted_count"`
-
-	// Schedule corresponds to the JSON schema field "schedule".
-	Schedule *string `json:"schedule,omitempty,omitzero"`
-
-	// SourceCounts corresponds to the JSON schema field "source_counts".
-	SourceCounts []NotebookDreamSourceCount `json:"source_counts"`
-
-	// Timezone corresponds to the JSON schema field "timezone".
-	Timezone *string `json:"timezone,omitempty,omitzero"`
-
-	// Top corresponds to the JSON schema field "top".
-	Top []NotebookDreamCandidate `json:"top"`
-}
-
 type NotebookEntry struct {
 	// Kind corresponds to the JSON schema field "kind".
 	Kind *string `json:"kind,omitempty,omitzero"`
@@ -2776,13 +2677,6 @@ type Response struct {
 
 	// Error corresponds to the JSON schema field "error".
 	Error *string `json:"error,omitempty,omitzero"`
-
-	// NotebookDreamRun corresponds to the JSON schema field "notebook_dream_run".
-	NotebookDreamRun *NotebookDreamRunResult `json:"notebook_dream_run,omitempty,omitzero"`
-
-	// NotebookDreamStatus corresponds to the JSON schema field
-	// "notebook_dream_status".
-	NotebookDreamStatus *NotebookDreamStatusResult `json:"notebook_dream_status,omitempty,omitzero"`
 
 	// NotebookEntries corresponds to the JSON schema field "notebook_entries".
 	NotebookEntries []NotebookEntry `json:"notebook_entries,omitempty,omitzero"`
