+++ internal/protocol/schema/main.tsp
@@ -660,18 +660,6 @@ model NotebookGuideMessage {
   session_id?: string; // when set, session_is_chief reflects this session's role
 }
 
-// Dreaming inspection (phase one): a deterministic, LLM-free harvest surfaced
-// read-only. notebook_dream_status returns a summary; notebook_dream_run returns
-// the candidate preview. Neither writes durable memory.
-model NotebookDreamStatusMessage {
-  cmd: "notebook_dream_status";
-}
-
-model NotebookDreamRunMessage {
-  cmd: "notebook_dream_run";
-  apply?: boolean; // reserved for the promote pass; harvest is preview-only for now
-}
-
 // Task status surface: list the durable task runner's records, and retry a
 // failed/dead one. state and kind are plain strings (not enums) so the on-disk
 // record stays self-describing and a new TypeSpec enum cannot trip the quicktype
@@ -1397,7 +1385,7 @@ model WorkspaceContextChangedMessage {
   updated_at: string;
 }
 
-// Broadcast when notebook content changes. origin is agent|dreaming|ui|external.
+// Broadcast when notebook content changes. origin is agent|ui|external.
 model NotebookChangedMessage {
   event: "notebook_changed";
   paths: string[];
@@ -2001,8 +1989,6 @@ model Response {
   notebook_read?: NotebookReadResult;
   notebook_write?: NotebookWriteResult;
   notebook_guide?: NotebookGuideResult;
-  notebook_dream_status?: NotebookDreamStatusResult;
-  notebook_dream_run?: NotebookDreamRunResult;
   notebook_tasks?: NotebookTask[];
 }
 
@@ -2082,57 +2068,6 @@ model NotebookGuideResult {
   guidance: string;
   root: string;
   session_is_chief: boolean;
-}
-
-// One deduplicated dreaming harvest candidate. occurrences is the number of
-// distinct source units carrying the fact; contexts are the distinct context
-// labels (workspaces, journal dates) — recurrence across contexts is the
-// promote phase's strongest durability signal. sources are resolvable
-// provenance pointers; snippet is word-bounded.
-model NotebookDreamCandidate {
-  signal_key: string;
-  source: string;
-  title?: string;
-  snippet: string;
-  occurrences: int32;
-  contexts: string[];
-  sources: string[];
-  first_seen?: string;
-  last_seen?: string;
-}
-
-model NotebookDreamSourceCount {
-  source: string;
-  count: int32;
-}
-
-// Summary of what a dream would consolidate now. enabled reflects the
-// notebook.dreaming.enabled gate; multi_context_count is the subset of
-// candidates seen in more than one distinct context. candidate_count is the live
-// preview (persisted candidates merged with a fresh harvest); persisted_count is
-// how many are durably saved on disk from prior scheduled runs. schedule/timezone
-// describe the cron; last_run_at/next_run_at bracket the scheduled runs.
-model NotebookDreamStatusResult {
-  enabled: boolean;
-  candidate_count: int32;
-  multi_context_count: int32;
-  persisted_count: int32;
-  source_counts: NotebookDreamSourceCount[];
-  top: NotebookDreamCandidate[];
-  schedule?: string;
-  timezone?: string;
-  last_run_at?: string;
-  next_run_at?: string;
-}
-
-// Result of notebook_dream_run. applied is false in the preview-only phase (the
-// run wrote nothing); candidates is the bounded harvested preview.
-model NotebookDreamRunResult {
-  applied: boolean;
-  candidate_count: int32;
-  multi_context_count: int32;
-  source_counts: NotebookDreamSourceCount[];
-  candidates: NotebookDreamCandidate[];
 }
 
 // One durable task-runner record, surfaced read-only for the task status panel.
